use PlainTextFormatter for channel search indexing
Extracts link text, document names, and user emails into
searchable content instead of stripping them with NullXmlFormatter.
Also consolidates the duplicate PlainTextFormatter from
model_notifications into mention_utils as a shared public type.
